cleaned up some more, and moved some outputWorker logic to seperate folder and package
