+++ .gitignore
@@ -19,7 +19,7 @@
 
 **/states.json
 **/routes.txt
-**/output*
+**/output.txt*
 **/performance*
 **/**.bin
 **/routes.json
